split v1 & v2 endpoints
