+++ classifai-core/src/main/java/ai/classifai/router/EndpointsV2.java
@@ -1,0 +1,240 @@
+package ai.classifai.router;
+
+import ai.classifai.database.portfolio.PortfolioDbQuery;
+import ai.classifai.loader.ProjectLoader;
+import ai.classifai.selector.filesystem.FileSystemStatus;
+import ai.classifai.selector.project.ProjectFolderSelector;
+import ai.classifai.util.ParamConfig;
+import ai.classifai.util.ProjectHandler;
+import ai.classifai.util.http.HTTPResponseHandler;
+import ai.classifai.util.message.ErrorCodes;
+import ai.classifai.util.message.ReplyHandler;
+import ai.classifai.util.type.AnnotationType;
+import io.vertx.core.Vertx;
+import io.vertx.core.eventbus.DeliveryOptions;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.RoutingContext;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class EndpointsV2 {
+
+    @Setter
+    private Vertx vertx;
+    public ProjectFolderSelector projectFolderSelector;
+
+    Utils util = new Utils();
+
+    /***
+     * change is_load state of a project to false
+     *
+     * PUT http://localhost:{port}/:annotation_type/projects/:project_name
+     */
+    public void closeProjectState(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+        String projectID = ProjectHandler.getProjectID(projectName, type.ordinal());
+
+        if(util.checkIfProjectNull(context, projectID, projectName)) return;
+
+        context.request().bodyHandler(h ->
+        {
+            try
+            {
+                JsonObject jsonObject = h.toJsonObject();
+
+                if(jsonObject.getString(ParamConfig.getStatusParam()).equals("closed"))
+                {
+                    ProjectHandler.getProjectLoader(projectID).setIsLoadedFrontEndToggle(Boolean.FALSE);
+                }
+                else
+                {
+                    throw new Exception("Request payload failed to satisfied the status of {\"status\": \"closed\"} for " + projectName + ". ");
+                }
+
+                HTTPResponseHandler.configureOK(context);
+
+            }
+            catch (Exception e)
+            {
+                HTTPResponseHandler.configureOK(context, ReplyHandler.reportUserDefinedError(e.getMessage()));
+
+            }
+        });
+    }
+
+    /***
+     * Star a project
+     *
+     * PUT http://localhost:{port}/:annotation_type/projects/:projectname/star
+     */
+    public void starProject(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+        String projectID = ProjectHandler.getProjectID(projectName, type.ordinal());
+
+        if(util.checkIfProjectNull(context, projectID, projectName)) return;
+
+        context.request().bodyHandler(h ->
+        {
+            JsonObject jsonObject = h.toJsonObject();
+
+            jsonObject.put(ParamConfig.getProjectIdParam(), projectID);
+
+            DeliveryOptions options = new DeliveryOptions().addHeader(ParamConfig.getActionKeyword(), PortfolioDbQuery.getStarProject());
+
+            vertx.eventBus().request(PortfolioDbQuery.getQueue(), jsonObject, options, reply ->
+            {
+                if(reply.succeeded())
+                {
+                    JsonObject response = (JsonObject) reply.result().body();
+
+                    HTTPResponseHandler.configureOK(context, response);
+                }
+            });
+        });
+    }
+
+    /**
+     * Create new project
+     * PUT http://localhost:{port}/v2/:annotation_type/newproject/:project_name
+     *
+     * Example:
+     * PUT http://localhost:{port}/v2/bndbox/newproject/helloworld
+     *
+     */
+    public void createV2Project(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+
+        projectFolderSelector.run(projectName, type);
+
+        HTTPResponseHandler.configureOK(context);
+    }
+
+    /**
+     * Reload v2 project
+     * PUT http://localhost:{port}/v2/:annotation_type/projects/:project_name/reload
+     *
+     * Example:
+     * PUT http://localhost:{port}/v2/bndbox/projects/helloworld/reload
+     *
+     */
+    public void reloadV2Project(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+
+        log.info("Reloading project: " + projectName + " of annotation type: " + type.name());
+
+        ProjectLoader loader = ProjectHandler.getProjectLoader(projectName, type);
+
+        if(util.checkIfProjectNull(context, loader, projectName)) return;
+
+        loader.setFileSystemStatus(FileSystemStatus.WINDOW_CLOSE_LOADING_FILES);
+
+        JsonObject jsonObject = new JsonObject().put(ParamConfig.getProjectIdParam(), loader.getProjectID());
+
+        DeliveryOptions reloadOpts = new DeliveryOptions().addHeader(ParamConfig.getActionKeyword(), PortfolioDbQuery.getReloadProject());
+
+        vertx.eventBus().request(PortfolioDbQuery.getQueue(), jsonObject, reloadOpts, fetch ->
+        {
+            JsonObject response = (JsonObject) fetch.result().body();
+
+            if (ReplyHandler.isReplyOk(response))
+            {
+                HTTPResponseHandler.configureOK(context);
+
+            } else
+            {
+                HTTPResponseHandler.configureOK(context, ReplyHandler.reportUserDefinedError("Failed to reload project " + projectName));
+            }
+        });
+    }
+
+    /**
+     * Get load status of project
+     * GET http://localhost:{port}/v2/:annotation_type/projects/:project_name/reloadstatus
+     *
+     * Example:
+     * GET http://localhost:{port}/v2/bndbox/projects/helloworld/reloadstatus
+     *
+     */
+    public void reloadV2ProjectStatus(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+
+        ProjectLoader loader = ProjectHandler.getProjectLoader(projectName, type);
+
+        if(util.checkIfProjectNull(context, loader, projectName));
+
+        FileSystemStatus fileSysStatus = loader.getFileSystemStatus();
+
+        JsonObject res = new JsonObject().put(ReplyHandler.getMessageKey(), fileSysStatus.ordinal());
+
+        if(fileSysStatus.equals(FileSystemStatus.WINDOW_CLOSE_DATABASE_UPDATING))
+        {
+            res.put(ParamConfig.getProgressMetadata(), loader.getProgressUpdate());
+        }
+        else if(fileSysStatus.equals(FileSystemStatus.WINDOW_CLOSE_DATABASE_UPDATED))
+        {
+            res.put(ParamConfig.getUuidAdditionListParam(), loader.getReloadAdditionList());
+            res.put(ParamConfig.getUuidDeletionListParam(), loader.getReloadDeletionList());
+        }
+        else if(fileSysStatus.equals(FileSystemStatus.DID_NOT_INITIATE))
+        {
+            res.put(ReplyHandler.getErrorCodeKey(), ErrorCodes.USER_DEFINED_ERROR.ordinal());
+            res.put(ReplyHandler.getErrorMesageKey(), "File / folder selection for project: " + projectName + " did not initiated");
+        }
+
+        HTTPResponseHandler.configureOK(context, res);
+    }
+
+    /***
+     * export a project to json
+     *
+     * PUT http://localhost:{port}/v2/:annotation_type/projects/:project_name/export
+     */
+    public void exportV2Project(RoutingContext context)
+    {
+        AnnotationType type = util.getAnnotationType(context.request().getParam(ParamConfig.getAnnotationTypeParam()));
+
+        String projectName = context.request().getParam(ParamConfig.getProjectNameParam());
+        String projectID = ProjectHandler.getProjectID(projectName, type.ordinal());
+
+        if(util.checkIfProjectNull(context, projectID, projectName)) return;
+
+        JsonObject request = new JsonObject()
+                .put(ParamConfig.getProjectIdParam(), projectID)
+                .put(ParamConfig.getAnnotationTypeParam(), type.ordinal());
+
+        DeliveryOptions options = new DeliveryOptions().addHeader(ParamConfig.getActionKeyword(), PortfolioDbQuery.getExportProject());
+
+        vertx.eventBus().request(PortfolioDbQuery.getQueue(), request, options, reply -> {
+
+            if (reply.succeeded()) {
+
+                JsonObject response = (JsonObject) reply.result().body();
+
+                HTTPResponseHandler.configureOK(context, response);
+
+            }
+            else
+            {
+                HTTPResponseHandler.configureOK(context, ReplyHandler.reportUserDefinedError("Export of project failed for " + projectName));
+
+            }
+        });
+
+    }
+}
